implemented review changes.
-mydir removed
-loader removed now all functionality in bsl_loader class.
-except replaced with except Exception
