+++ src/sas/sascalc/file_converter/loader.py
@@ -6,72 +6,8 @@ import numpy as np
 class Loader():
 
     def __init__(self, filename, frames, pixels, rasters, swap_bytes):
-        self.filename = filename
-        self.n_frames = frames
-        self.n_pixels = pixels
-        self.n_rasters = rasters
-        self.swap_bytes = swap_bytes
-        self.frame = np.int()
 
-    #File to load.
-    @property
-    def filename(self):
-        return self._filename
 
-    @filename.setter
-    def filename(self, filename):
-        _filename = self.try_convert(str, filename)
-        self._filename = _filename
-
-    #Number of frames in the file.
-    @property
-    def n_frames(self):
-        return self._n_frames
-
-    @n_frames.setter
-    def n_frames(self, n_frames):
-        _n_frames = self.try_convert(int, n_frames)
-        self._n_frames = _n_frames
-
-    #Frame to load.
-    @property
-    def frame(self):
-        return self._frame
-
-    @frame.setter
-    def frame(self, frame):
-        _frame = self.try_convert(int, frame)
-        self._frame = _frame
-
-    #Number of pixels in the file.
-    @property
-    def n_pixels(self):
-        return self._n_pixels
-
-    @n_pixels.setter
-    def n_pixels(self, n_pixels):
-        _n_pixels = self.try_convert(int, n_pixels)
-        self._n_pixels = _n_pixels
-
-    #Number of rasters in the file (int)
-    @property
-    def n_rasters(self):
-        return self._n_rasters
-
-    @n_rasters.setter
-    def n_rasters(self, n_rasters):
-        _n_rasters = self.try_convert(int, n_rasters)
-        self._n_rasters = _n_rasters
-
-    #Whether or not the bytes are in reverse order (int)
-    @property
-    def swap_bytes(self):
-        return self._swap_bytes
-
-    @swap_bytes.setter
-    def swap_bytes(self, swap_bytes):
-        _swap_bytes = self.try_convert(bool, swap_bytes)
-        self._swap_bytes = _swap_bytes
 
 
     def try_convert(self, type, value):
